+++ contracts/libraries/muon/LibMuonSettlement.sol
@@ -17,7 +17,7 @@ library LibMuonSettlement {
 		bytes memory encodedData;
 		bytes memory encodedPartyBPrices;
 		uint256[] memory nonces = new uint256[](settleSig.quotesSettlementsData.length);
-		for (uint8 i = 0; i < settleSig.quotesSettlementsData.length; i++) {
+		for (uint256 i = 0; i < settleSig.quotesSettlementsData.length; i++) {
 			nonces[i] = AccountStorage.layout().partyBNonces[QuoteStorage.layout().quotes[settleSig.quotesSettlementsData[i].quoteId].partyB][partyA];
 			encodedData = abi.encodePacked(
 				encodedData,  // Append the previously encoded data
@@ -25,7 +25,7 @@ library LibMuonSettlement {
 				settleSig.quotesSettlementsData[i].currentPrice
 			);
 		}
-		for (uint8 i = 0; i < settleSig.partyBPriceSigs.length; i++) {
+		for (uint256 i = 0; i < settleSig.partyBPriceSigs.length; i++) {
 			encodedPartyBPrices = abi.encodePacked(
 				encodedPartyBPrices,
 				settleSig.partyBPriceSigs[i].quoteIds,
